minor validataion

+++ playlists_url_download/combine_urls.sh
@@ -1,7 +1,12 @@
 #!/bin/bash
 
 all_merged_urls='./output/all_merged_urls.txt' 
-rm "$all_merged_urls"
+
+if [ -f "$all_merged_urls" ];then
+    rm "$all_merged_urls"
+fi
+
+exit 
 
 #fileNames=`ls -rt *_result.txt`
 fileNames=`ls -rt ./output/vds_*.txt`
